all the functionality of librarian is done

+++ src/main/java/com/library/management/service/impl/GenreServiceImpl.java
@@ -1,5 +1,6 @@
 package com.library.management.service.impl;
 
+import com.library.management.dao.BookDAO;
 import com.library.management.dao.GenreDAO;
 import com.library.management.model.Genre;
 import com.library.management.service.GenreService;
@@ -15,6 +16,7 @@ import java.util.List;
 public class GenreServiceImpl implements GenreService {
 
     private final GenreDAO genreDAO;
+    private final BookDAO bookDAO;
 
     @Override
     @Transactional(readOnly = true)
@@ -29,6 +31,13 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public void delete(Genre genre) {
+        long inUse = bookDAO.countByGenreId(genre.getId());
+        if (inUse > 0) {
+            throw new IllegalStateException(
+                    String.format("Cannot delete genre “%s”: it’s still used by %d book(s).",
+                            genre.getName(), inUse)
+            );
+        }
         genreDAO.delete(genre);
     }
 }
